requirements.txt
fastapi
uvicorn
pydantic
torch
Pillow